+++ deploy/02_deploy_streams.ts
@@ -44,17 +44,18 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     hre.deployments,
     accounts[0],
     accounts[0],
-    "ERC721StakingEmissionStream",
+    "ERC721LockedStakingEmissionStream",
     [
       {
         ticketToken: "0x0000000000000000000000000000000000000000",
         lockedUntilTimestamp: 0,
-        minStakingLockTime: 0,
+        minStakingDuration: 0,
         maxStakingTotalDurations: 0,
         emissionRate: 0,
         emissionTimeUnit: 0,
         emissionStart: 0,
         emissionEnd: 0,
+        totalTickets: 0,
         claimLockedUntil: 0,
       },
     ]